fix initial implementation of renderToPipeableStream method
