+++ packages/subapp-react/lib/framework-lib.js
@@ -10,11 +10,32 @@ const { default: AppContext } = require("../dist/node/app-context");
 const ReactRedux = optionalRequire("react-redux", { default: {} });
 const { Provider } = ReactRedux;
 const ReactRouterDom = optionalRequire("react-router-dom");
-
+const { Stream } = require("stream");
 class FrameworkLib {
   constructor(ref) {
     this.ref = ref;
     this._prepared = false;
+  }
+
+  getStreamWritable() {
+    const writable = new Stream.PassThrough();
+    writable.setEncoding("utf8");
+    const output = { result: "", error: undefined };
+    writable.on("data", chunk => {
+      output.result += chunk;
+    });
+    writable.on("error", error => {
+      output.error = error;
+    });
+    const completed = new Promise(resolve => {
+      writable.on("finish", () => {
+        resolve();
+      });
+      writable.on("error", () => {
+        resolve();
+      });
+    });
+    return { writable, completed, output };
   }
 
   async handlePrepare() {
@@ -57,7 +78,7 @@ class FrameworkLib {
     return "";
   }
 
-  renderTo(element, options) {
+  async renderTo(element, options) {
     const { subAppServer } = this.ref;
 
     if (typeof subAppServer.renderer === "function") {
@@ -65,7 +86,11 @@ class FrameworkLib {
     }
 
     if (options.suspenseSsr || options.useStream || options.hydrateServerData) {
-      return ReactDOMServer.renderToPipeableStream(element);
+      const { writable, output, completed } = this.getStreamWritable();
+      const { pipe } = await ReactDOMServer.renderToPipeableStream(element);
+      pipe(writable);
+      await completed;
+      return output.result;
     } else {
       return ReactDOMServer.renderToStaticMarkup(element);
     }
